+++ src/test/kotlin/me/bossm0n5t3r/leetcode/palindromicSubstrings/PalindromicSubstringsTest.kt
@@ -1,11 +1,11 @@
-package palindromicSubstrings
+package me.bossm0n5t3r.leetcode.palindromicSubstrings
 
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
 import kotlin.test.assertEquals
 
 internal class PalindromicSubstringsTest {
-    private val palindromicSubstrings = PalindromicSubstrings()
+    private val palindromicSubstrings = PalindromicSubstrings.Solution()
 
     private data class PalindromicSubstringsTestData(
         val s: String,
